+++ test/test_example.py
@@ -12,6 +12,7 @@ from example.main import Modis
 import cumulus_process.s3 as s3
 from cumulus_process import Process, helpers
 
+os.environ['LOCALSTACK_HOST'] = 'localhost'
 
 class TestExample(unittest.TestCase):
     """Test the example implementation of the Process class"""